+++ CHANGES.md
@@ -1,5 +1,7 @@
 # Version 0.6.4
 
+- Added function plotSpikePSD to plot power spectral density of spiking data
+
 - Added plotShape to plot 3D morphology of cell and synapse locations
 
 - Added option to skip batch sims if output file already exists